requirements.txt
streamlit
langchain
transformers
sentence-transformers
chromadb
python-dotenv
requests
torch
